LayoutTests/imported/w3c/web-platform-tests/css/css-shapes/parsing/shape-outside-valid-position.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Shapes Module Level 1: parsing shape-outside with valid position values</title>
<link rel="author" title="Eric Willigers" href="mailto:ericwilligers@chromium.org">
<link rel="help" href="https://drafts.csswg.org/css-values-4/#typedef-position">
<link rel="help" href="https://github.com/w3c/csswg-drafts/issues/8695">
<meta name="assert" content="shape-outside positions support the full '<position>' grammar.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value("shape-outside", "circle(at 10% 20%)");
test_valid_value("shape-outside", "circle(at 10%)", "circle(at 10% center)");
test_valid_value("shape-outside", "circle(at 20% 30px)");
test_valid_value("shape-outside", "circle(at 30px center)", "circle(at 30px center)");
test_valid_value("shape-outside", "circle(at 40px top)", "circle(at 40px top)");
test_valid_value("shape-outside", "circle(at bottom 10% right 20%)", "circle(at right 20% bottom 10%)");
test_valid_value("shape-outside", "circle(at bottom calc(10%) right calc(20%))", "circle(at right calc(20%) bottom calc(10%))");
test_valid_value("shape-outside", "circle(at bottom right)", "circle(at right bottom)");
test_valid_value("shape-outside", "circle(at center 10em)");
test_valid_value("shape-outside", "circle(at center 50px)", "circle(at center 50px)");
test_valid_value("shape-outside", "circle(at center bottom)", "circle(at center bottom)");
test_valid_value("shape-outside", "circle(at center center)", "circle(at center center)");
test_valid_value("shape-outside", "circle(at center left)", "circle(at left center)");
test_valid_value("shape-outside", "circle(at center)", "circle(at center center)");
test_valid_value("shape-outside", "circle(at left bottom)", "circle(at left bottom)");
test_valid_value("shape-outside", "circle(at left center)", "circle(at left center)");
test_valid_value("shape-outside", "circle(at left)", "circle(at left center)");
test_valid_value("shape-outside", "circle(at right 30% top 60px)", "circle(at right 30% top 60px)");
test_valid_value("shape-outside", "circle(at right 40%)", "circle(at right 40%)");
test_valid_value("shape-outside", "circle(at right)", "circle(at right center)");
test_valid_value("shape-outside", "circle(at top center)", "circle(at center top)");
test_valid_value("shape-outside", "circle(at top)", "circle(at center top)");

test_valid_value("shape-outside", "ellipse(at 10% 20%)");
test_valid_value("shape-outside", "ellipse(at 10%)", "ellipse(at 10% center)");
test_valid_value("shape-outside", "ellipse(at 20% 30px)");
test_valid_value("shape-outside", "ellipse(at 30px center)", "ellipse(at 30px center)");
test_valid_value("shape-outside", "ellipse(at 40px top)", "ellipse(at 40px top)");
test_valid_value("shape-outside", "ellipse(at bottom 10% right 20%)", "ellipse(at right 20% bottom 10%)");
test_valid_value("shape-outside", "ellipse(at bottom calc(10%) right calc(20%))", "ellipse(at right calc(20%) bottom calc(10%))");
test_valid_value("shape-outside", "ellipse(at bottom right)", "ellipse(at right bottom)");
test_valid_value("shape-outside", "ellipse(at center 10em)");
test_valid_value("shape-outside", "ellipse(at center 50px)", "ellipse(at center 50px)");
test_valid_value("shape-outside", "ellipse(at center bottom)", "ellipse(at center bottom)");
test_valid_value("shape-outside", "ellipse(at center center)", "ellipse(at center center)");
test_valid_value("shape-outside", "ellipse(at center left)", "ellipse(at left center)");
test_valid_value("shape-outside", "ellipse(at center)", "ellipse(at center center)");
test_valid_value("shape-outside", "ellipse(at left bottom)", "ellipse(at left bottom)");
test_valid_value("shape-outside", "ellipse(at left center)", "ellipse(at left center)");
test_valid_value("shape-outside", "ellipse(at left)", "ellipse(at left center)");
test_valid_value("shape-outside", "ellipse(at right 30% top 60px)", "ellipse(at right 30% top 60px)");
test_valid_value("shape-outside", "ellipse(at right 40%)", "ellipse(at right 40%)");
test_valid_value("shape-outside", "ellipse(at right)", "ellipse(at right center)");
test_valid_value("shape-outside", "ellipse(at top center)", "ellipse(at center top)");
test_valid_value("shape-outside", "ellipse(at top)", "ellipse(at center top)");

</script>
</body>
</html>
