LayoutTests/imported/w3c/web-platform-tests/css/css-shapes/parsing/shape-outside-valid-position-expected.txt


FAIL e.style['shape-outside'] = "circle(at 10%)" should set the property value assert_equals: serialization should be canonical expected "circle(at 10% 50%)" but got "circle(at 10% center)"
PASS e.style['shape-outside'] = "circle(at 20% 30px)" should set the property value
FAIL e.style['shape-outside'] = "circle(at 30px center)" should set the property value assert_equals: serialization should be canonical expected "circle(at 30px 50%)" but got "circle(at 30px center)"
FAIL e.style['shape-outside'] = "circle(at 40px top)" should set the property value assert_equals: serialization should be canonical expected "circle(at 40px 0%)" but got "circle(at 40px top)"
FAIL e.style['shape-outside'] = "circle(at bottom 10% right 20%)" should set the property value assert_equals: serialization should be canonical expected "circle(at calc(80%) calc(90%))" but got "circle(at right 20% bottom 10%)"
FAIL e.style['shape-outside'] = "circle(at bottom right)" should set the property value assert_equals: serialization should be canonical expected "circle(at 100% 100%)" but got "circle(at right bottom)"
FAIL e.style['shape-outside'] = "circle(at center)" should set the property value assert_equals: serialization should be canonical expected "circle(at 50% 50%)" but got "circle(at center center)"
FAIL e.style['shape-outside'] = "circle(at center 50px)" should set the property value assert_equals: serialization should be canonical expected "circle(at 50% 50px)" but got "circle(at center 50px)"
FAIL e.style['shape-outside'] = "circle(at center bottom)" should set the property value assert_equals: serialization should be canonical expected "circle(at 50% 100%)" but got "circle(at center bottom)"
FAIL e.style['shape-outside'] = "circle(at center center)" should set the property value assert_equals: serialization should be canonical expected "circle(at 50% 50%)" but got "circle(at center center)"
FAIL e.style['shape-outside'] = "circle(at center left)" should set the property value assert_equals: serialization should be canonical expected "circle(at 0% 50%)" but got "circle(at left center)"
FAIL e.style['shape-outside'] = "circle(at left)" should set the property value assert_equals: serialization should be canonical expected "circle(at 0% 50%)" but got "circle(at left center)"
FAIL e.style['shape-outside'] = "circle(at left bottom)" should set the property value assert_equals: serialization should be canonical expected "circle(at 0% 100%)" but got "circle(at left bottom)"
FAIL e.style['shape-outside'] = "circle(at left center)" should set the property value assert_equals: serialization should be canonical expected "circle(at 0% 50%)" but got "circle(at left center)"
FAIL e.style['shape-outside'] = "circle(at right 40%)" should set the property value assert_equals: serialization should be canonical expected "circle(at 100% 40%)" but got "circle(at right 40%)"
FAIL e.style['shape-outside'] = "circle(at right 30% top 60px)" should set the property value assert_equals: serialization should be canonical expected "circle(at calc(70%) 60px)" but got "circle(at right 30% top 60px)"
FAIL e.style['shape-outside'] = "circle(at top)" should set the property value assert_equals: serialization should be canonical expected "circle(at 50% 0%)" but got "circle(at center top)"
FAIL e.style['shape-outside'] = "circle(at top center)" should set the property value assert_equals: serialization should be canonical expected "circle(at 50% 0%)" but got "circle(at center top)"
FAIL e.style['shape-outside'] = "circle(at top 0% right calc(10% * sign(1em - 1px)))" should set the property value assert_equals: serialization should be canonical expected "circle(at calc(100% - (10% * sign(1em - 1px))) 0%)" but got "circle(at right calc(10% * sign(1em - 1px)) top 0%)"
FAIL e.style['shape-outside'] = "circle(at top 0% right calc(10% * sibling-index()))" should set the property value assert_equals: serialization should be canonical expected "circle(at calc(100% - (10% * sibling-index())) 0%)" but got "circle(at right calc(10% * sibling-index()) top 0%)"
